Use the GAT feature for HKT support
This needs the nightly toolchain to run, but it's much cleaner than the
implementation through helper traits, and it's closer to what HKTs are actually
are.

Also, force all wrapped values for functors to implement Clone so that the
return values are always copies of the input and the functions remain pure.